// extensions/erdos-python/src/client/common/application/contextKeys.ts
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.

export enum ExtensionContextKey {
    showInstallPythonTile = 'showInstallPythonTile',
    HasFailedTests = 'hasFailedTests',
    RefreshingTests = 'refreshingTests',
    IsJupyterInstalled = 'isJupyterInstalled',
}