// extensions/erdos-python/src/client/common/application/walkThroughs.ts
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.

export enum PythonWelcome {
    name = 'pythonWelcome',
    windowsInstallId = 'python.installPythonWin8',
    linuxInstallId = 'python.installPythonLinux',
    macOSInstallId = 'python.installPythonMac',
}